% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{diff2SVR}
\alias{diff2SVR}
\title{Reverse \code{SVR2diff}}
\usage{
diff2SVR(diff, len_merge = 0)
}
\arguments{
\item{len_merge}{Merge SVRs that are no more \code{len_merge}'nt far.}
}
\description{
Reverse \code{\link{SVR2diff}}
}
