% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize DESSERT}
\usage{
init(r, qc_ctrl = FALSE, ncores = "auto")
}
\arguments{
\item{r}{A list of reactivities of two conditions.}

\item{qc_ctrl}{Quality control settings. See "Details".}

\item{ncores}{Number of cores for parallel computation. Automatically determined by default.}
}
\value{
Initialized reactivities.
}
\description{
Initialize with optional quality control.
}
\details{
Set \code{qc_ctrl = FALSE} to skip QC. Otherwise, \code{qc_ctrl} should be a list
with components \code{n_trim_tail} and \code{bad_sites}. \code{n_trim_tail} nucleotide
positions in the tail of each transcript will be trimmed. Nucleotide positions in
\code{bad_sites} will be removed.

Transcripts are cut into segments with length not exceeding 100 nt.
}
