% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize reactivities}
\usage{
normalize(r, fake0 = FALSE, plotfile = FALSE, ncores = "auto", verbose = TRUE)
}
\arguments{
\item{r}{Initialized reactivities from \code{\link{init}}.}

\item{fake0}{A logical variable indicating whether there are a substantial number
of fake 0's in reactivities (\code{FALSE} by default). Fake 0's arises when SP
experiments fail to measure the structure of part of the nucleotide positions.}

\item{plotfile}{File name of plot of details. Set to \code{FALSE} (default) to skip.}

\item{ncores}{Number of cores for parallel computation. Automatically determined by default.}

\item{verbose}{A logical variable indicating whether to print details (\code{TRUE} by default).}
}
\value{
A list containing the following components:
\item{\code{r}}{Normalized reactivities.}
\item{\code{fit}}{The fitted model for the normalization.}
}
\description{
Remove systematic bias before differential analysis. Skip this step
if the user is convinced of the comparability between reactivities.
}
\details{
Users should inspect normalized reactivities carefully.
}
