% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nscan}
\alias{nscan}
\title{Return the number of nucleotide positions of a transcript that are scanned}
\usage{
nscan(scan)
}
\description{
Return the number of nucleotide positions of a transcript that are scanned
}
