% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rcut2rf}
\alias{rcut2rf}
\title{Convert rcut of a transcript into a tibble}
\usage{
rcut2rf(r_cut)
}
\arguments{
\item{r_cut}{Reactivities that are cut into a list.}
}
\value{
A tibble with columns \code{pos} indicating position.
}
\description{
Convert rcut of a transcript into a tibble
}
