% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan}
\alias{scan}
\title{Scan to identify structurally variable regions (SVRs)}
\usage{
scan(r, seed, alpha = 0.05, N = 1000, ncores = "auto")
}
\arguments{
\item{r}{Initialized reactivity from \code{\link{init}} or normalized reactivity
from \code{\link{normalize}}.}

\item{seed}{Random seed for reproducible results.}

\item{alpha}{Significance level.}

\item{N}{Number of replications in Monte Carlo sampling.}

\item{ncores}{Number of cores to use for parallel computation. Automatically determined
by default.}
}
\value{
A list with each element corresponding to a transcript. \code{svr} is
SVRs with significance score \code{P}. \code{scan} includes the start and stop positions
of all scanned segments.
}
\description{
This function detects structurally variable regions in nucleotide resolution and
controls family-wise error rate.
}
