% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate reactivity}
\usage{
simulate(
  seq,
  dir_ViennaRNA,
  types,
  ncores = "auto",
  plot_cor = FALSE,
  clear = TRUE
)
}
\arguments{
\item{seq}{List of RNA sequences.}

\item{dir_ViennaRNA}{Directory of ViennaRNA software. See Details.}

\item{types}{Vector of reactivity types ('icSHAPE', 'Cordero' and 'Sokusd' supported).}

\item{ncores}{Number of cores for parallel computation. Automatically determined by default.}

\item{plot_cor}{Logical value indicating whether to plot within/between group correlations.}

\item{clear}{Logical value indicating whether to clear intermediate files.}
}
\value{
A list containing the following components:
\item{\code{r}}{Simulated reactivity.}
\item{\code{stru}}{A list of simulated structural conformations.}
\item{\code{truth}}{A list of vectors indicating positions of structurally variable regions.}
\item{\code{w}}{Weights of dominant conformations.}
}
\description{
Given RNA sequences, this function simulates reactivity accounting for
multiple structural conformations and dependencies between adjacent nucleotide positions.
}
\details{
Users need to have ViennaRNA software installed to simulate structural
conformations from sequences.
}
