% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SVR2diff}
\alias{SVR2diff}
\title{Convert SVRs of a transcript into a status vector}
\usage{
SVR2diff(SVR, n)
}
\arguments{
\item{n}{Length of the transcript.}
}
\description{
The status vector indicates whether each nucleotide position is in SVRs.
}
